requirements.txt
django==5.2.1
django-ninja==1.4.1
django-ninja-jwt==5.3.7
httpx==0.28.1
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pyright==1.11.401
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
pytest-mock==3.14.0
PyYAML==6.0.1
ruff==0.11.11
uvicorn==0.34.2
